+++ HISTORY.rst
@@ -1,6 +1,12 @@
 History
 =======
 
+0.7.3
+-----
+
+* Hide confusing password field from user change form
+* Custom admin menu items can be added now using project settings
+
 0.7.2
 -----